z8000: add userland C library, fix coff2aout data relocations
Build the FUZIX C library for Z8001 segmented mode:
- Syscall stub generator with type-aware arg loading (16-bit int
  vs 32-bit pointer stack layout differs on z8k)
- crt0 and crt0nostdio with VDSO 0x20 space, correct argc/argv
  offsets (exec stores argc as 32-bit via uputl), and ___main stub
- setjmp/longjmp (Z8001 CALL pushes segment at lower address,
  offset at higher — opposite of what I initially assumed)
- Linker wrapper (ldz8k) using mktemp to avoid input/output
  file collision with ld -r

Fix coff2aout to process .data section relocations. Previously
only .text relocations were handled, leaving all pointers in
initialized data (stdio FILE buffers, shell's brkincr, etc.)
unrelocated — pointing to segment 0 (kernel space) instead of
the user segment. This caused EFAULT on write() from printf
and zero brkincr in the shell's allocator.

Other fixes in this commit:
- Fork wrapper clears r2-r5 for _fork(flags=0, addr=NULL)
- Fork wrapper saves/restores r8-r13 on user stack so child
  gets correct callee-saved values (68000 port approach)
- forkreturn no longer clears r8-r13 callee-saved registers
- Shell round() macro: int→intptr_t to prevent 32-bit pointer
  truncation on z8k (int is 16-bit)
- Target/rules.z8000: use ldz8k linker wrapper
- Application Makefiles for z8000 target (sh, V7 cmd, util)

Tested: printf, argc/argv, setjmp/longjmp all pass on MAME M20.

+++ Applications/V7/cmd/sh/defs.h
@@ -107,7 +107,7 @@ extern char **sh_setenv(void);
 
 
 #define attrib(n,f)	(n->namflg |= f)
-#define round(a,b)	(((int)((ADR(a)+b)-1))&~((b)-1))
+#define round(a,b)	(((intptr_t)((ADR(a)+b)-1))&~((b)-1))
 #define closepipe(x)	(close(x[INPIPE]), close(x[OTPIPE]))
 #define eq(a,b)		(cf(a,b)==0)
 #define max(a,b)	((a)>(b)?(a):(b))
